## Description
Using C++ Build Insights in VS2019 I recorded a build
and looked at time spent parsing header files. This PR
adds some of the slow headers to pch.

Rough numbers on before after on my machine
Debug:
  - Before: 00:11:35
  - After: 00:07:58

 RelWithDebInfo
   - Before: 00:21:48
   - After: 00:17:15

 measured using
 msbuild kodi.sln -t:Clean -P:Configuration=RelWithDebInfo
 msbuild kodi.sln -t:Build -P:Configuration=RelWithDebInfo

|Included Path|Count|Exclusive Duration (s)|
|---------|-------------|-------------------------|
|C:\Program Files (x86)\Microsoft Visual Studio\2019\Community\VC\Tools\MSVC\14.25.28610\include\zmmintrin.h|977|107,579000000|
|C:\Program Files (x86)\Microsoft Visual Studio\2019\Community\VC\Tools\MSVC\14.25.28610\include\ppltasks.h|507|97,202000000|
|C:\Program Files (x86)\Microsoft Visual Studio\2019\Community\VC\Tools\MSVC\14.25.28610\include\chrono|943|72,215000000|
|C:\code\kodi\project\BuildDependencies\x64\include\fmt\format.h|947|58,281000000|
|C:\code\kodi\xbmc\FileItem.h|562|50,436000000|
|C:\code\kodi\xbmc\addons\addoninfo\AddonInfo.h|609|47,643000000|
|C:\code\kodi\project\BuildDependencies\x64\include\fmt\core.h|947|35,932000000|
|C:\Program Files (x86)\Microsoft Visual Studio\2019\Community\VC\Tools\MSVC\14.25.28610\include\algorithm|1064|32,647000000|
|C:\code\kodi\xbmc\windowing\GraphicContext.h|450|25,386000000|
|C:\code\kodi\xbmc\addons\addoninfo\AddonExtensions.h|610|23,582000000|
|C:\Program Files (x86)\Windows Kits\10\Include\10.0.18362.0\um\DirectXMathVector.inl|58|22,396000000|
|C:\code\kodi\xbmc\guilib\GUIWindow.h|329|21,691000000|
|C:\code\kodi\xbmc\utils\Variant.h|624|20,434000000|
|C:\Program Files (x86)\Microsoft Visual Studio\2019\Community\VC\Tools\MSVC\14.25.28610\include\mutex|936|19,288000000|
|C:\code\kodi\xbmc\settings\lib\Setting.h|176|18,965000000|
|C:\code\kodi\xbmc\guilib\GUIWindowManager.h|247|18,292000000|
|C:\Program Files (x86)\Microsoft Visual Studio\2019\Community\VC\Tools\MSVC\14.25.28610\include\atomic|633|17,678000000|
|C:\code\kodi\xbmc\windowing\WinSystem.h|459|17,489000000|
|C:\code\kodi\project\BuildDependencies\x64\include\tinyxml.h|449|15,892000000|
|C:\Program Files (x86)\Microsoft Visual Studio\2019\Community\VC\Tools\MSVC\14.25.28610\include\functional|569|15,486000000|
|C:\Program Files (x86)\Microsoft Visual Studio\2019\Community\VC\Tools\MSVC\14.25.28610\include\future|485|14,856000000|
|C:\Program Files (x86)\Microsoft Visual Studio\2019\Community\VC\Tools\MSVC\14.25.28610\include\list|787|14,314000000|
|C:\Program Files (x86)\Microsoft Visual Studio\2019\Community\VC\Tools\MSVC\14.25.28610\include\intrin.h|977|14,184000000|
|C:\code\kodi\xbmc\ServiceManager.h|130|14,184000000|

## Motivation and Context
<!--- Why is this change required? What problem does it solve? -->
<!--- If it fixes an open issue, please link to the issue here -->

## How Has This Been Tested?
<!--- Please describe in detail how you tested your change -->
<!--- Include details of your testing environment, and the tests you ran to -->
<!--- see how your change affects other areas of the code, etc -->

## Screenshots (if appropriate):

## Types of change
<!--- What type of change does your code introduce? Put an `x` in all the boxes that apply like this: [X] -->
- [ ] **Bug fix** (non-breaking change which fixes an issue)
- [ ] **Clean up** (non-breaking change which removes non-working, unmaintained functionality)
- [x] **Improvement** (non-breaking change which improves existing functionality)
- [ ] **New feature** (non-breaking change which adds functionality)
- [ ] **Breaking change** (fix or feature that will cause existing functionality to change)
- [ ] **Cosmetic change** (non-breaking change that doesn't touch code)
- [ ] **None of the above** (please explain below)

## Checklist:
<!--- Go over all the following points, and put an `X` in all the boxes that apply like this: [X] -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're here to help! -->
- [x] My code follows the **[Code Guidelines](https://github.com/xbmc/xbmc/blob/master/docs/CODE_GUIDELINES.md)** of this project 
- [ ] My change requires a change to the documentation, either Doxygen or wiki
- [ ] I have updated the documentation accordingly
- [x] I have read the **[Contributing](https://github.com/xbmc/xbmc/blob/master/docs/CONTRIBUTING.md)** document
- [ ] I have added tests to cover my change
- [x] All new and existing tests passed