@neo1973 2020-04-13
Just something to keep in mind for the future, native support for precompiled headers was added to CMake in 3.16:  https://cmake.org/cmake/help/latest/command/target_precompile_headers.html
Once the minimum required CMake version is raised it would probably be beneficial to switch to that method to speed up compilation on all platforms.