@garbear 2020-04-14
The amount of coffee I don't drink waiting for kodi to compile is coming out of your paycheck. Nice timing stats.