@garbear 2020-04-14
When the clang commit comes first, git blame is more beautiful with relevant commit messages appearing instead of clang format lemons.